fix v2.3.2+66: 修正 3 個分析器警告/錯誤
- sticker_canvas.dart:428 - int 轉 double (argument_type_not_assignable)
- sticker_canvas.dart:1 - 移除多餘的 dart:typed_data import
- editor_screen.dart:355 - 移除未使用的 _StyleBar class

+++ lib/features/editor/widgets/sticker_canvas.dart
@@ -1,5 +1,3 @@
-import 'dart:typed_data';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -425,7 +423,7 @@ class _TextSelectionWidget extends StatelessWidget {
 
         // 旋轉控制點（文字上方：24px 圓圈 + 20px 連接線，4px 間距）
         Positioned(
-          top: -(24 + 20 + 4), // 頂部留出 handle 高度 + 4px 間距
+          top: -(24 + 20 + 4).toDouble(), // 頂部留出 handle 高度 + 4px 間距
           child: const _RotationHandle(),
         ),
